+++ cloud/oauth-server/src/main/kotlin/pers/acp/admin/oauth/controller/open/inner/OpenInnerUserController.kt
@@ -4,7 +4,6 @@ import io.swagger.annotations.*
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
-import org.springframework.security.oauth2.provider.OAuth2Authentication
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
 import pers.acp.admin.common.base.BaseController